Adjust bundle models to avoid unnecessary reads
Bundle models are now instantiated with a list of filenames rather than
the contents of the bundle. This avoids unnecessary reading of the
original assets.

Bundle models also have a new way of computing their hash: via the
filenames and the modified times, rather than the contents. Again, this
change aims to have the original assets read only when necessary.